pyproject.toml
[tool.poetry]
name = "jbt_drone"
version = "0.1.0"
description = ""
authors = ["André Felipe Dias <andref.dias@gmail.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
fastapi = "*"
hypercorn = "*"
loguru = "*"
orjson = "*"
PyJWT = "*"
python = "^3.10"
python-dotenv = "*"
python-multipart = "*"
uvloop = "*"

[tool.poetry.dev-dependencies]
asgi-lifespan = "*"
alt-pytest-asyncio = "*"
blue = "*"
bandit = "*"
flake8 = "*"
flake8-debugger = "*"
flake8-pytest-style = "*"
httpx = "*"
isort = "*"
mypy = "*"
pep8-naming = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
pyupgrade = "*"

[tool.isort]
profile = "black"
line_length = 82

[tool.blue]
line-length = 82

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
